requirements.txt
scanpy
numpy
pandas
matplotlib
seaborn
scikit-learn
umap-learn
torch
python-dateutil
